Connecting property getter and setter.
I'm a little annoyed by this approach but I felt like anymore would be
sort of over-engineering the a solution. I could have refactored to use
everything to use a a recursive AST-building like approach (returning
a Method and assigning that to Property.Getter, etc..) or added
StartPropertySetter, etc... This was just easier... Later on if there
is more of a technical demand I'll change how this is done. What's
important is that the model is getting constructed and it's not too ugly.