Got usages recording!
This introduces a more complex Usage hierarchy and an abstract factory
method on Node that creates the Usage for that given sub-type. So
the Type.Usage() call will return a TypeUsage. We use that in UsageSet
to build up a set of usages.

Weirdness in this commit is the GenericParameterTypeName. Now that I have
some generics in the analyzed code I had problems with them. They show
up as a TypeName (which makes sense since they can be used as one) which
was throwing everything off completely... This won't really stay but it's
a good start towards something better (a sort of null object pattern)
Right now they all get recorded in the None assembly, hehe.