Working on IInNode interface and implementations.
I'm trying to think of the best way to organize and get the querying
functionality working. I'm thinking I need a concept of querying as
well as filtering and selecting.

Selecting is choosing something from a set of things I've queried for and
filtered? If that makes sense... Eh, I'm thinking this over still.