Trying to come up with a model for the Query and Select API.
I'm trying to find the best implementation that gives me the most value for
the least amount of work. Isn't that what we're all doing? I'm trying
to avoid doing some crazy SQL-like work... although some database
theory would probably do me some good here.

Instead of trying to build a Condition hierarchy can I could build a
single Condition for I decided it'd be easier to just build a hard coded
Query class that has the conditions for the various syntactic items. Source
code won't be changing that much in the future (it'll have methods, types,
namespaces, etc...) so there's no real value in getting too crazy with
this. From there it's a matter of selection and filtering more. We're
effectively doing the FROM and the JOINs with Query right now.