Refactoring how direct uses are stored.
I'm trying to keep them as close to the thing that's using them as
possible and merging them to create a useful behavior, rather than
making assumptions now. For example if people wanted to get just
the uses of a property (and its methods) the design will support that. Or
say the uses of all fields or or all methods but not properties, etc...