Indirect usages stuff.
Got the self appearing in the indirect uses and started on some
specifications to ensure we don't regress. Ran into some weirdness with
the ShouldContainOnly that I'm going to talk to Aaron about. For now
I have my own implementation that has uglier messages but behaves the
way I expected.