Refactoring towards better CQS.
Broke the creation of members into their own methods. We still do the
FindOrCreate thing though for now. Moved a bunch of stuff off of
Assembly and into the repository. There's really two things going on
here, because once we get an Assembly, FindOrCreate* is NOT part
of the API.