New Pending node handling.
Whenever we try and get information that we don't have about a node
because that node is pending we throw an exception. This is a hard
signal to the user that they just can't get those details until they
load the assembly that the node is in.

It's not the prettiest implementation of this feature, I'll agree.