Working on getting external assemblies loading.
This introduces a PendingTypeLoader that iterates over assemblies that
have pending types and tries to load their AssemblyDefinition's. That's
pretty much it right now. We try and do GAC loads and then we iterate
over paths we got the other assemblies from, using .dll and .exe to find
them physically on disk (if GAC fails)