Visiting assemblies that have pending types.
This takes forever now because we do no filtering on the types we're
visiting. I've also fixed several new issues that popped up as
stack overflows. This required some stacks and some checks.

There's also a VisitationOptions that we can use to reject visiting
method bodies and will also help with filtering later.