New TopDownVisitor.
I was struggling with the built in visitation mechanism inside of Cecil
because it's difficult to wrap the entire visit of a Type and exclude
things. The reason being that the TypeDefinition.Accept will hit the
visitor and then continue to visit the members and such. I need the ability
to skip the members easily and there's just no easy way to build that
in. This approach makes it very clear what is being visited and when and
wasn't all that hard to adapt.