Type.DirectlyUsedByMembers works better now.
It'll go over methods, properties, fields, etc.. and gather up all of
those type references that are on them. Like return types, parameter
types, field types, etc... We combine those with the direct uses of
the methods themselves.

Also a bunch of specifications, most of which pass except for some border
cases dealing with the types themselves and the external dependencies.